spatialite-gui: fix broken build
Remove hacks that were breaking the build. Ostensibly they fixed
something at one time but they don't seem to work in either stdenv or
superenv now.

Also list some deps that are already in the dependency tree, but are
linked into the binary so they should be repeated here.

Fixes #24101.